plant_tracker_back/plants/serializers.py
from rest_framework import serializers
from .models import Plant

class PlantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Plant
        fields = ['name', 'slug', 'room', 'last_time_watered']